Explain the main purpose of this script
A reason why this script is maybe better for demos than for everyday
deploying is that the script asks you which port to use and then checks
whether it's available rather than checking which ports are available
and asking you which one to use. Maybe this will change in the future.

One thing that could cause a problem for demos is the fact that this
script may display the full contents of a local .env file to help the
user choose what a remote one should have. I plan on changing this so
that the user won't potentially show secrets on their screen while
screen sharing.